style(feature/DomainLayer): Improve test readability
