+++ SwiftCleanBootcamp/DataTests/Helpers/Factories.swift
@@ -1,0 +1,16 @@
+//
+//  Factories.swift
+//  DataTests
+//
+//  Created by Willian Guedes on 22/10/22.
+//
+
+import Foundation
+
+func makeInvalidData() -> Data {
+    return Data("invalid_data".utf8)
+}
+
+func makeURL() -> URL {
+    return URL(string: "https://www.google.com.br")!
+}
